Day 3

+++ OOPS/counter/Program.cs
@@ -1,10 +1,26 @@
 ﻿using System.IO;
 Console.WriteLine("Counter Demo");
 
-int counter = int.Parse(File.ReadLines("counter.txt").LastOrDefault() ?? "0");
+StreamWriter myPen;
+FileStream counterFile;
 
-Console.WriteLine($"You have run this program {counter + 1} times.");
+if (File.Exists("counter.txt"))
+{
+    int counter = int.Parse(File.ReadLines("counter.txt").LastOrDefault() ?? "0");
 
-StreamWriter myPen = new StreamWriter("counter.txt");
-myPen.WriteLine(counter + 1);
-myPen.Close();
+    Console.WriteLine($"You have run this program {counter + 1} times.");
+
+    myPen = new StreamWriter("counter.txt");
+    myPen.WriteLine(counter + 1);
+    myPen.Close();
+}
+else
+{
+    counterFile = new FileStream("counter.txt", FileMode.Create, FileAccess.Write);
+
+    Console.WriteLine($"You have run this program 1 time.");
+
+    myPen = new StreamWriter("counter.txt");
+    myPen.WriteLine("1");
+    myPen.Close();
+}
